Move fix_pymol_saving to pymol interface and map to pymol config
