+++ enzy_htp/chemical/residue.py
@@ -560,46 +560,6 @@ The key is the residue after the mutation."""
 
 # yapf: enable
 
-"""A map to convert messed up PyMOL atom names to a more conventional name for the Structure class."""
-PYMOL_TO_ATOM_MAPPER: Dict[str, str] = {
-    "1HB": "HB1",
-    "2HB": "HB2",
-    "3HB": "HB3",
-    "1HG": "HG1",
-    "2HG": "HG2",
-    "3HG": "HG3",
-    "1HD": "HD1",
-    "2HD": "HD2",
-    "3HD": "HD3",
-    "1HE": "HE1",
-    "2HE": "HE2",
-    "3HE": "HE3",
-    "2HH": "HH2",
-    "1HZ": "HZ1",
-    "2HZ": "HZ2",
-    "3HZ": "HZ3",
-    "1HH1": "HH11",
-    "2HH1": "HH12",
-    "1HH2": "HH21",
-    "2HH2": "HH22",
-    "1HD2": "HD21",
-    "2HD2": "HD22",
-    "1HE2": "HE21",
-    "2HE2": "HE22",
-    "1HG2": "HG21",
-    "2HG2": "HG22",
-    "3HG2": "HG23",
-    "1HG1": "HG11",
-    "2HG1": "HG12",
-    "3HG1": "HG13",
-    "1HD1": "HD11",
-    "2HD1": "HD12",
-    "3HD1": "HD13",
-    "1HD2": "HD21",
-    "2HD2": "HD22",
-    "3HD2": "HD23",
-}
-
 def convert_to_three_letter(one_letter: str) -> str:
     """Converts a one letter amino acid name to a three letter. If supplied code is invalid, raises an enzy_htp.core.InvalidResidueCode() exception."""
     if len(one_letter) != 1:
