+++ test/_interface/data/Stru_TRP.pdb
