+++ test/_interface/data/pymol_TRP.pdb
